Merge branch 0.98

+++ hbase-common/src/main/java/org/apache/hadoop/hbase/util/CoprocessorClassLoader.java
@@ -18,6 +18,7 @@
 package org.apache.hadoop.hbase.util;
 
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.net.URL;
@@ -39,6 +40,9 @@ import org.apache.hadoop.hbase.classification.InterfaceAudience;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.fs.FileStatus;
+import org.apache.hadoop.fs.FileUtil;
+import org.apache.hadoop.hbase.classification.InterfaceAudience;
 import org.apache.hadoop.io.IOUtils;
 
 import com.google.common.base.Preconditions;
@@ -97,7 +101,6 @@ public class CoprocessorClassLoader extends ClassLoaderBase {
   private static final String[] CLASS_PREFIX_EXEMPTIONS = new String[] {
     // Java standard library:
     "com.sun.",
-    "launcher.",
     "java.",
     "javax.",
     "org.ietf",
@@ -110,7 +113,19 @@ public class CoprocessorClassLoader extends ClassLoaderBase {
     "org.apache.log4j",
     "com.hadoop",
     // Hadoop/HBase/ZK:
-    "org.apache.hadoop",
+    "org.apache.hadoop.security",
+    "org.apache.hadoop.HadoopIllegalArgumentException",
+    "org.apache.hadoop.conf",
+    "org.apache.hadoop.fs",
+    "org.apache.hadoop.http",
+    "org.apache.hadoop.io",
+    "org.apache.hadoop.ipc",
+    "org.apache.hadoop.metrics",
+    "org.apache.hadoop.metrics2",
+    "org.apache.hadoop.net",
+    "org.apache.hadoop.util",
+    "org.apache.hadoop.hdfs",
+    "org.apache.hadoop.hbase",
     "org.apache.zookeeper",
   };
 
@@ -144,7 +159,7 @@ public class CoprocessorClassLoader extends ClassLoaderBase {
     super(parent);
   }
 
-  private void init(Path path, String pathPrefix,
+  private void init(Path pathPattern, String pathPrefix,
       Configuration conf) throws IOException {
     // Copy the jar to the local filesystem
     String parentDirStr =
@@ -162,31 +177,53 @@ public class CoprocessorClassLoader extends ClassLoaderBase {
       }
     }
 
-    FileSystem fs = path.getFileSystem(conf);
-    File dst = new File(parentDirStr, "." + pathPrefix + "."
-      + path.getName() + "." + System.currentTimeMillis() + ".jar");
-    fs.copyToLocalFile(path, new Path(dst.toString()));
-    dst.deleteOnExit();
+    FileSystem fs = pathPattern.getFileSystem(conf);
+    Path pathPattern1 = fs.isDirectory(pathPattern) ?
+      new Path(pathPattern, "*.jar") : pathPattern;  // append "*.jar" if a directory is specified
+    FileStatus[] fileStatuses = fs.globStatus(pathPattern1);  // return all files that match the pattern
+    if (fileStatuses == null || fileStatuses.length == 0) {  // if no one matches
+      throw new FileNotFoundException(pathPattern1.toString());
+    } else {
+      boolean validFileEncountered = false;
+      for (Path path : FileUtil.stat2Paths(fileStatuses)) {  // for each file that match the pattern
+        if (fs.isFile(path)) {  // only process files, skip for directories
+          File dst = new File(parentDirStr, "." + pathPrefix + "."
+            + path.getName() + "." + System.currentTimeMillis() + ".jar");
+          fs.copyToLocalFile(path, new Path(dst.toString()));
+          dst.deleteOnExit();
 
-    addURL(dst.getCanonicalFile().toURI().toURL());
+          addURL(dst.getCanonicalFile().toURI().toURL());
 
-    JarFile jarFile = new JarFile(dst.toString());
-    try {
-      Enumeration<JarEntry> entries = jarFile.entries();
-      while (entries.hasMoreElements()) {
-        JarEntry entry = entries.nextElement();
-        Matcher m = libJarPattern.matcher(entry.getName());
-        if (m.matches()) {
-          File file = new File(parentDirStr, "." + pathPrefix + "."
-            + path.getName() + "." + System.currentTimeMillis() + "." + m.group(1));
-          IOUtils.copyBytes(jarFile.getInputStream(entry),
-            new FileOutputStream(file), conf, true);
-          file.deleteOnExit();
-          addURL(file.toURI().toURL());
+          JarFile jarFile = new JarFile(dst.toString());
+          try {
+            Enumeration<JarEntry> entries = jarFile.entries();  // get entries inside a jar file
+            while (entries.hasMoreElements()) {
+              JarEntry entry = entries.nextElement();
+              Matcher m = libJarPattern.matcher(entry.getName());
+              if (m.matches()) {
+                File file = new File(parentDirStr, "." + pathPrefix + "."
+                  + path.getName() + "." + System.currentTimeMillis() + "." + m.group(1));
+                FileOutputStream outStream = new FileOutputStream(file);
+                try {
+                  IOUtils.copyBytes(jarFile.getInputStream(entry),
+                    outStream, conf, true);
+                } finally {
+                  outStream.close();
+                }
+                file.deleteOnExit();
+                addURL(file.toURI().toURL());
+              }
+            }
+          } finally {
+            jarFile.close();
+          }
+
+          validFileEncountered = true;  // Set to true when encountering a file
         }
       }
-    } finally {
-      jarFile.close();
+      if (validFileEncountered == false) {  // all items returned by globStatus() are directories
+        throw new FileNotFoundException("No file found matching " + pathPattern1.toString());
+      }
     }
   }
 
@@ -227,7 +264,7 @@ public class CoprocessorClassLoader extends ClassLoaderBase {
       return cl;
     }
 
-    if (!pathStr.endsWith(".jar")) {
+    if (path.getFileSystem(conf).isFile(path) && !pathStr.endsWith(".jar")) {
       throw new IOException(pathStr + ": not a jar file?");
     }
 
@@ -266,8 +303,13 @@ public class CoprocessorClassLoader extends ClassLoaderBase {
   @Override
   public Class<?> loadClass(String name)
       throws ClassNotFoundException {
+    return loadClass(name, null);
+  }
+
+  public Class<?> loadClass(String name, String[] includedClassPrefixes)
+      throws ClassNotFoundException {
     // Delegate to the parent immediately if this class is exempt
-    if (isClassExempt(name)) {
+    if (isClassExempt(name, includedClassPrefixes)) {
       if (LOG.isDebugEnabled()) {
         LOG.debug("Skipping exempt class " + name +
             " - delegating directly to parent");
@@ -346,7 +388,14 @@ public class CoprocessorClassLoader extends ClassLoaderBase {
    * @return true if the class should *not* be loaded by this ClassLoader;
    * false otherwise.
    */
-  protected boolean isClassExempt(String name) {
+  protected boolean isClassExempt(String name, String[] includedClassPrefixes) {
+    if (includedClassPrefixes != null) {
+      for (String clsName : includedClassPrefixes) {
+        if (name.startsWith(clsName)) {
+          return false;
+        }
+      }
+    }
     for (String exemptPrefix : CLASS_PREFIX_EXEMPTIONS) {
       if (name.startsWith(exemptPrefix)) {
         return true;
